bootloader: menu is now displayed in all cases
regardless of whether there is an image or not in flash
we display the menu. This is to handle the case of missing
or corrupt firmware.

Removed some obsolete test code that is in the way.

The bootloader still needs to put up the menu *before*
attempting anything fancy that could conceivably fail.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>

git-svn-id: http://www.ecosforge.net/ecosforge/trunk/nios2ecos@322 2333b21a-1c27-0410-8132-afba56cb5048