ethermac: MAC address is now specified by fixed location in flash
The fixed location is configured in the ecos.ecc file.

Why a fixed location?

Perhaps a fn with a weak symbol that the application can
the override would be better? How would multiple
interfaces be handled? Perhaps force application to use
POSIX API to modify MAC address that way? Perhaps a more
clever "cleaner" way later on? This is a pragmatic first
step though.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>

git-svn-id: http://www.ecosforge.net/ecosforge/trunk/nios2ecos@328 2333b21a-1c27-0410-8132-afba56cb5048